Move overflow test to live among other overflow tests
